+++ resolwe_bio/processes/microarray/methylation_arrays.py
@@ -28,13 +28,15 @@ class MethylationArraySesame(ProcessBio):
     some basic statistics, such as mean beta, fraction of
     (un)methylated, GCT, predicted ethnicity, gender and age.
     Methylation data file holds betas, mvals and pvals for probe ids.
-    In addition, Ensembl IDs and HGNC gene symbol names are provided.
+    In addition, Ensembl IDs and HGNC gene symbol names are provided,
+    along with chromosome and start/end positions of CpG sites
+    (1-based).
     """
 
     slug = "methylation-array-sesame"
     name = "Methylation analysis (SeSAMe)"
     process_type = "data:methylation:sesame"
-    version = "1.2.1"
+    version = "1.2.2"
     category = "Methylation arrays"
     data_name = 'SeSAMe array ({{ idat_file.red_channel.file|default("?") }})'
     scheduling_class = SchedulingClass.BATCH